# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: RedGain membership referral platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /payments:
    post:
      operationId: submitPayment
      tags: [payments]
      summary: Submit a payment proof
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "201":
          description: Payment submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentRecord"
...
  /admin/users/{id}:
    patch:
      operationId: adminUpdateUser
      tags: [admin]
      summary: Update user status or role (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: adminDeleteUser
      tags: [admin]
      summary: Delete a user and their payments (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Cannot delete admin account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/payments/{id}:
    patch:
      operationId: adminReviewPayment
      tags: [admin]
      summary: Approve or reject a payment (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentReview"
      responses:
        "200":
          description: Updated payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminPayment"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
...
    PaymentInput:
      type: object
      required: [proofText, amount]
      properties:
        proofText:
          type: string
        amount:
          type: number
        proofImageUrl:
          type: ["string", "null"]
        proofImageUrls:
          type: array
          items:
            type: string
    PaymentRecord:
      type: object
      required: [id, userId, amount, paymentType, status, proofText, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        amount:
          type: number
        paymentType:
          type: string
          enum: [initial, renewal]
        status:
          type: string
          enum: [pending, approved, rejected]
        proofText:
          type: string
        txHash:
          type: ["string", "null"]
          description: Blockchain transaction hash (for crypto payments)
        proofImageUrl:
          type: ["string", "null"]
        proofImageUrls:
          type: array
          items:
            type: string
        createdAt:
          type: string
        reviewedAt:
          type: ["string", "null"]
    AdminUser:
      type: object
      required: [id, name, email, role, accountStatus, referralCode, totalReferrals, activeReferrals, joinedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        whatsappUrl:
          type: ["string", "null"]
        role:
          type: string
        accountStatus:
          type: string
          enum: [pending, active, paused, lost]
        referralCode:
          type: string
        referrerName:
          type: ["string", "null"]
        referrerId:
          type: ["integer", "null"]
        totalReferrals:
          type: integer
        activeReferrals:
          type: integer
        membershipStartedAt:
          type: ["string", "null"]
        membershipTimerStartedAt:
          type: ["string", "null"]
        membershipExpiresAt:
          type: ["string", "null"]
        gracePeriodEndsAt:
          type: ["string", "null"]
        daysRemaining:
          type: ["integer", "null"]
        inGracePeriod:
          type: boolean
        timerStarted:
          type: boolean
        joinedAt:
          type: string
    AdminUserUpdate:
      type: object
      properties:
        accountStatus:
          type: string
          enum: [pending, active, paused, lost]
        role:
          type: string
          enum: [user, admin]
    AdminPayment:
      type: object
      required: [id, userId, userName, userEmail, amount, paymentType, status, proofText, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        userEmail:
          type: string
        userPhone:
          type: ["string", "null"]
        amount:
          type: number
        paymentType:
          type: string
          enum: [initial, renewal]
        status:
          type: string
          enum: [pending, approved, rejected]
        proofText:
          type: string
        proofImageUrl:
          type: ["string", "null"]
        proofImageUrls:
          type: array
          items:
            type: string
        createdAt:
          type: string
        reviewedAt:
          type: ["string", "null"]
    PaymentReview:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [approved, rejected]
        note:
          type: ["string", "null"]
...